refactor(style-editor): Simplify form handling and improve state management
- Refactored the style editor form component to utilize a dedicated form builder service, enhancing form creation and management.
- Removed unused imports and cleaned up code for better readability.
- Updated methods for extracting and updating style properties to streamline functionality and improve performance.
- Ensured consistent handling of form values during rollback operations.

These changes enhance the maintainability and efficiency of the style editor, providing a more robust user experience.
